Further fixes

+++ api/src/main/java/au/org/aodn/nrmn/restapi/repository/SpeciesWithAttributesRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpeciesWithAttributesRepository extends JpaRepository<SpeciesWithAttributes, Long>, JpaSpecificationExecutor<SpeciesWithAttributes> {
+public interface SpeciesWithAttributesRepository extends JpaRepository<SpeciesWithAttributes, Integer>, JpaSpecificationExecutor<SpeciesWithAttributes> {
 }
